+++ src/main/java/com/openclassrooms/chatopapi/repository/UserRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.openclassrooms.chatopapi.model.User;
 
-
 public interface UserRepository extends CrudRepository<User, Integer> {
 	Optional<User> findByEmail(String email);
-	Optional<User> findById(Integer id);
+
+	Optional<User> findById(Long id);
 }
